requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
constants==0.6.0
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
geographiclib==2.0
geopy==2.3.0
idna==3.4
imageio-ffmpeg==0.4.8
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.25.0
option==2.1.0
protobuf==3.20.3
py-cord==2.4.1
pycord==0.1.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
yarl==1.9.2

pycountry~=22.3.5